# openapi.yaml
...
openapi: 3.1.0
info:
  title: Enkrypt Backend API
  version: 1.0.0
paths:
...
  /backups/{publicKey}:
    get:
      operationId: GetUserBackups
      parameters:
        - $ref: '#/components/parameters/PathPublicKey'
        - $ref: '#/components/parameters/QuerySignature'
      responses:
        200:
          $ref: '#/components/responses/GetUserBackupsSuccess'
  /backups/{publicKey}/users/{userId}:
    post:
      operationId: CreateUserBackup
      parameters:
        - $ref: '#/components/parameters/PathPublicKey'
        - $ref: '#/components/parameters/PathUserId'
        - $ref: '#/components/parameters/QuerySignature'
      requestBody:
        $ref: '#/components/requestBodies/CreateUserBackup'
      responses:
        200:
          $ref: '#/components/responses/PostUserBackupSuccess'
    delete:
      operationId: DeleteUserBackup
      parameters:
        - $ref: '#/components/parameters/PathPublicKey'
        - $ref: '#/components/parameters/PathUserId'
        - $ref: '#/components/parameters/QuerySignature'
      responses:
        200:
          $ref: '#/components/responses/DeleteUserBackupSuccess'
components:
  parameters:
    PathPublicKey:
      name: publicKey
      in: path
      # required: true # OpenAPI 3.0 property not compatible with JSON schema
      schema:
        $ref: '#/components/schemas/PublicKey'
    PathUserId:
      name: userId
      in: path
      # required: true # OpenAPI 3.0 property not compatible with JSON schema
      schema:
        $ref: '#/components/schemas/UserId'
    QuerySignature:
      name: signature
      in: query
      # required: true # OpenAPI 3.0 property not compatible with JSON schema
      schema:
        $ref: '#/components/schemas/ByteString'
  requestBodies:
    CreateUserBackup:
      content:
        application/json:
          $ref: '#/components/schemas/CreateUserBackupRequest'
  responses:
...
    GetUserBackupsSuccess:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GetUserBackupsResponse'
    PostUserBackupSuccess:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PostUserBackupResponse'
    DeleteUserBackupSuccess:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DeleteUserBackupResponse'
  schemas:
    ByteString:
      type: string
      pattern: ^0x([0-9a-fA-F]{2})*$
...
    Bytes64:
      type: string
      pattern: ^0x[0-9a-fA-F]{128}$
...
    UUID:
      type: string
      pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
    PublicKey:
      $ref: '#/components/schemas/Bytes64'
    UserId:
      $ref: '#/components/schemas/UUID'
...
    CreateUserBackupRequest:
      type: object
      required:
        - payload
      properties:
        payload:
          $ref: '#/components/schemas/ByteString'
    PostUserBackupResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    DeleteUserBackupResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    GetUserBackupsResponse:
      type: object
      required:
        - backups
      properties:
        backups:
          type: array
          items:
            $ref: '#/components/schemas/Backup'
    Backup:
      type: object
      required:
        - updatedAt
        - userId
        - payload
      properties:
        updatedAt:
          type: string
          format: date-time
        userId:
          type: string
        payload:
          $ref: '#/components/schemas/ByteString'